feat(ANT): PrimitiveUpdateFromPast
So weapons don't work very well with prim updates since they are mostly event based and snapshots don't have anything important.
To solve this we use repvars (specifically uint8 Proxy_ItemData) and then we call PrimitiveUpdateFromPast which passes in the previous Base constructed from snaps
For items, we just copy all the values from the previous tick and then use the revars to set input data and then call simulate over World()->GetDeltaSeconds().
To prevent time / order desync from messing up sim, we can store pending values, ie for astro rifle we store a PendingShotCount on the AAstroRifle which is the change in uin8 ItemData value, then as long as PendingShotCount > 0, bWantFire is true, and after simming, if FLAG_Fire is set, PendingShotCount--. This basically guarantees that the number of shots are same on all machines and they are space appropriety (since FireCooldownTime still applies).
For reload we just use Proxy_PendingReload to store reload until CanReload()